--- docs/src/app/api/feedback/lib/error.ts
@@ -1,24 +1,0 @@
-import { ErrorWithMessage } from "./types";
-
-function isErrorWithMessage(error: unknown): error is ErrorWithMessage {
-  return (
-    typeof error === "object" &&
-    error !== null &&
-    "message" in error &&
-    typeof (error as Record<string, unknown>).message === "string"
-  );
-}
-
-function toErrorWithMessage(maybeError: unknown): ErrorWithMessage {
-  if (isErrorWithMessage(maybeError)) return maybeError;
-
-  try {
-    return new Error(JSON.stringify(maybeError));
-  } catch {
-    return new Error(String(maybeError));
-  }
-}
-
-export function getErrorMessage(error: unknown) {
-  return toErrorWithMessage(error).message;
-}
